pyproject.toml
[tool.poetry]
name = "roz"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
boto3 = "^1.28.0"
opencv-python = "^4.8.0"
numpy = "^1.24.0"
matplotlib = "^3.7.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
